make splash screen wait for api calls
This also makes it so the splash has a bundle of API data in which the
API call methods can chuck their data. The bundle is then packed into
the intent to start the next activity and has been verified to having
correctly packed the data.

+++ app/src/main/java/com/example/timetomeet/retrofit/Endpoint.java
@@ -26,7 +26,7 @@ public interface Endpoint {
   Call<User> signUp(@Body User user);
 
   @GET("city/{cityId}/")
-  Call<City> getCityById(@Path("cityId") int cityId);
+  Call<City> getCityById(@Path("cityId") long cityId);
 
   @GET("paymentalternative/")
   Call<List<PaymentAlternative>> getPaymentAlternatives();
